Moved BracketContext to a context folder. Progress circle bar set up for Round Progress
