+++ services/frontend/src/App.scss
@@ -1,3 +1,5 @@
+@use 'stylesheets/progress.scss';
+
 .App {
   text-align: center;
 }
