+++ services/frontend/src/pages/ArtistPage.jsx
@@ -4,7 +4,7 @@ import {
 
 import BackToTop from '../components/BackToTop';
 import TopTracks from '../components/TopTracks';
-import BracketContext from '../components/BracketContext';
+import BracketContext from '../context/BracketContext';
 import ArtistPageForm from '../components/ArtistPageForm';
 import ArtistName from '../components/ArtistName';
 import ErrorBoundary from '../boundary/ErrorBoundary';
